+++ resources/js/components/RoomTabFiles.vue
@@ -326,7 +326,11 @@ const props = defineProps({
   },
 });
 
-const emit = defineEmits(["invalidRoomAuthToken", "guestsNotAllowed"]);
+const emit = defineEmits([
+  "invalidRoomAuthToken",
+  "requireCode",
+  "guestsNotAllowed",
+]);
 
 const api = useApi();
 const userPermissions = useUserPermissions();
@@ -422,7 +426,7 @@ function loadData(page = null) {
           error.response.status === env.HTTP_FORBIDDEN &&
           error.response.data.message === HTTP_ROOM_REQUIRE_CODE
         ) {
-          return emit("invalidRoomAuthToken");
+          return emit("requireCode");
         }
       }
       api.error(error, { redirectOnUnauthenticated: false });
@@ -457,7 +461,7 @@ function handleFileErrorMessages(event) {
     emit("invalidRoomAuthToken");
   } else if (event.data.type === HTTP_ROOM_REQUIRE_CODE) {
     // Forbidden, require access code
-    emit("invalidRoomAuthToken");
+    emit("requireCode");
   } else if (event.data.type === HTTP_FORBIDDEN) {
     // Forbidden, not allowed to view file
     toast.error(t("rooms.flash.file_forbidden"));