+++ resources/js/components/RoomTabSection.vue
@@ -85,6 +85,7 @@
           :room-auth-token="roomAuthToken"
           :room="props.room"
           @invalid-room-auth-token="$emit('invalidRoomAuthToken')"
+          @require-code="$emit('requireCode')"
           @guests-not-allowed="$emit('guestsNotAllowed')"
           @settings-changed="$emit('settingsChanged')"
         />
@@ -146,7 +147,12 @@ import { useUrlSearchParams } from "@vueuse/core";
 import { useSettingsStore } from "../stores/settings.js";
 import RoomTabStreaming from "./RoomTabStreaming.vue";
 
-defineEmits(["invalidRoomAuthToken", "guestsNotAllowed", "settingsChanged"]);
+defineEmits([
+  "invalidRoomAuthToken",
+  "requireCode",
+  "guestsNotAllowed",
+  "settingsChanged",
+]);
 
 const props = defineProps({
   room: {